+++ views/choose.html
@@ -1,0 +1,119 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <!-- Bootstrap -->
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/css/bootstrap.min.css"
+        integrity="sha384-TX8t27EcRE3e/ihU7zmQxVncDAy5uIKz4rEkgIXeMed4M0jlfIDPvg6uqKI2xXr2" crossorigin="anonymous"
+        type="text/css">
+    <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js"
+        integrity="sha384-DfXdz2htPH0lsSSs5nCTpuj/zy4C+OGpamoFVy38MVBnE+IbbVYUew+OrCXaRkfj"
+        crossorigin="anonymous"></script>
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/js/bootstrap.bundle.min.js"
+        integrity="sha384-ho+j7jyWK8fNQe+A12Hb8AhRq26LrZ/JpcUGGOn+Y7RsweNrtN/tE3MoK7ZeZDyx"
+        crossorigin="anonymous"></script>
+    <!-- Personal stylesheet -->
+    <link rel="stylesheet" href='css/style.min.css' type="text/css">
+    <link rel="stylesheet" href='css/nav.min.css' type="text/css">
+    <link rel="stylesheet" href="css/choose.min.css" type="text/css">
+    <title>Play-Gen</title>
+</head>
+
+<body>
+    <div class="container-fluid">
+        <div class="row nav-row">
+            <nav class="navbar navbar-expand-lg navbar-dark">
+                <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarTogglerDemo03"
+                    aria-controls="navbarTogglerDemo03" aria-expanded="false" aria-label="Toggle navigation">
+                    <svg width="1em" height="1em" viewBox="0 0 16 16" class="bi bi-arrow-bar-down" fill="currentColor"
+                        xmlns="http://www.w3.org/2000/svg">
+                        <path fill-rule="evenodd"
+                            d="M1 3.5a.5.5 0 0 1 .5-.5h13a.5.5 0 0 1 0 1h-13a.5.5 0 0 1-.5-.5zM8 6a.5.5 0 0 1 .5.5v5.793l2.146-2.147a.5.5 0 0 1 .708.708l-3 3a.5.5 0 0 1-.708 0l-3-3a.5.5 0 0 1 .708-.708L7.5 12.293V6.5A.5.5 0 0 1 8 6z" />
+                    </svg>
+                </button>
+                <a class="navbar-brand" href="#">Play-Gen</a>
+                <div class="collapse navbar-collapse" id="navbarTogglerDemo03">
+                    <ul class="navbar-nav mr-auto mt-2 mt-lg-0">
+                        {% if user %}
+                        <li class="nav-item active">
+                            <a class="nav-link" href="/">Home <span class="sr-only">(current)</span></a>
+                        </li>
+                        <li class="nav-item">
+                            <a class="nav-link" href="/rec">Playlist Creator</a>
+                        </li>
+                        <li class="nav-item">
+                            <a class="nav-link" href="/user-input">User Inputs</a>
+                        </li>
+                        <li class="nav-item"><a href="https://github.com/Mir-Khan/Play-Gen" class="nav-link">Github</a>
+                        </li>
+                        <li class="nav-item"><a href="https://www.mir-khan.com/" target="_blank"
+                                class="nav-link">Resume</a></li>
+                        {% else %}
+                        <li class="nav-item active">
+                            <a class="nav-link" href="/">Home <span class="sr-only">(current)</span></a>
+                        </li>
+                        <li class="nav-item"><a href="https://github.com/Mir-Khan/Play-Gen" class="nav-link">Github</a>
+                        </li>
+                        <li class="nav-item"><a href="https://www.mir-khan.com/" target="_blank"
+                                class="nav-link">Resume</a></li>
+                        {% endif %}
+                    </ul>
+                </div>
+            </nav>
+        </div>
+        <div class="row choose-row">
+            <div class="col text-center">
+                <form action="/rec" method="GET">
+                    <button type="submit" class="choose-button">
+                        <svg viewBox="0 0 16 16" class="bi bi-arrow-left-circle" xmlns="http://www.w3.org/2000/svg">
+                            <path fill-rule="evenodd"
+                                d="M8 15A7 7 0 1 0 8 1a7 7 0 0 0 0 14zm0 1A8 8 0 1 0 8 0a8 8 0 0 0 0 16z" />
+                            <path fill-rule="evenodd"
+                                d="M12 8a.5.5 0 0 1-.5.5H5.707l2.147 2.146a.5.5 0 0 1-.708.708l-3-3a.5.5 0 0 1 0-.708l3-3a.5.5 0 1 1 .708.708L5.707 7.5H11.5a.5.5 0 0 1 .5.5z" />
+                        </svg>
+                        <br>
+                        <p>Automatically generated playlists.<br>(Anyone can use this!)</p>
+                    </button>
+                </form>
+            </div>
+            <div class="col text-center">
+                <form action="/user-input" method="GET">
+                    <button type="submit" class="choose-button">
+                        <svg viewBox="0 0 16 16" class="bi bi-arrow-right-circle" xmlns="http://www.w3.org/2000/svg">
+                            <path fill-rule="evenodd"
+                                d="M8 15A7 7 0 1 0 8 1a7 7 0 0 0 0 14zm0 1A8 8 0 1 0 8 0a8 8 0 0 0 0 16z" />
+                            <path fill-rule="evenodd"
+                                d="M4 8a.5.5 0 0 0 .5.5h5.793l-2.147 2.146a.5.5 0 0 0 .708.708l3-3a.5.5 0 0 0 0-.708l-3-3a.5.5 0 1 0-.708.708L10.293 7.5H4.5A.5.5 0 0 0 4 8z" />
+                        </svg>
+                        <br>
+                        <p>User-input playlists.<br>(Recommended for newer accounts!)</p>
+                    </button>
+                </form>
+            </div>
+        </div>
+        <div class="row text-center justify-content-center logout">
+            <form action="/logout" method="GET">
+                <div>
+                    <button type="submit" class="btn"><span>Logout</span></button>
+                </div>
+            </form>
+        </div>
+    </div>
+    <footer class="footer">
+        <div class="fluid-container text-center">
+            <p>Made by Mir Khan</p>
+            {% if user %}
+            <p><a href="/">Home</a> | <a href="/rec">Playlist Creator</a> | <a href="/user-input">User
+                    Inputs</a> | <a href="https://github.com/Mir-Khan/Play-Gen" target="_blank">Github</a> | <a
+                    href="https://mir-khan.com" target="_blank">Resume</a></p>
+            {% else %}
+            <p><a href="/">Home</a> | <a href="https://github.com/Mir-Khan/Play-Gen" target="_blank">Github</a> | <a
+                    href="https://mir-khan.com" target="_blank">Resume</a></p>
+            {% endif %}
+        </div>
+    </footer>
+</body>
+
+</html>
