--- views/how-to.html
@@ -1,120 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <!-- Bootstrap -->
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/css/bootstrap.min.css"
-        integrity="sha384-TX8t27EcRE3e/ihU7zmQxVncDAy5uIKz4rEkgIXeMed4M0jlfIDPvg6uqKI2xXr2" crossorigin="anonymous"
-        type="text/css">
-    <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js"
-        integrity="sha384-DfXdz2htPH0lsSSs5nCTpuj/zy4C+OGpamoFVy38MVBnE+IbbVYUew+OrCXaRkfj"
-        crossorigin="anonymous"></script>
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/js/bootstrap.bundle.min.js"
-        integrity="sha384-ho+j7jyWK8fNQe+A12Hb8AhRq26LrZ/JpcUGGOn+Y7RsweNrtN/tE3MoK7ZeZDyx"
-        crossorigin="anonymous"></script>
-    <!-- Personal stylesheet -->
-    <link rel="stylesheet" href='css/style.min.css' type="text/css">
-    <title>Play-Gen</title>
-</head>
-
-<body>
-    <div class="container-fluid">
-        <div class="row nav-row">
-            <nav class="navbar navbar-expand-lg navbar-dark">
-                <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarTogglerDemo03"
-                    aria-controls="navbarTogglerDemo03" aria-expanded="false" aria-label="Toggle navigation">
-                    <svg width="1em" height="1em" viewBox="0 0 16 16" class="bi bi-arrow-bar-down" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
-						<path fill-rule="evenodd" d="M1 3.5a.5.5 0 0 1 .5-.5h13a.5.5 0 0 1 0 1h-13a.5.5 0 0 1-.5-.5zM8 6a.5.5 0 0 1 .5.5v5.793l2.146-2.147a.5.5 0 0 1 .708.708l-3 3a.5.5 0 0 1-.708 0l-3-3a.5.5 0 0 1 .708-.708L7.5 12.293V6.5A.5.5 0 0 1 8 6z"/>
-					  </svg>
-                </button>
-                <a class="navbar-brand" href="#">Playgen</a>
-
-                <div class="collapse navbar-collapse" id="navbarTogglerDemo03">
-                    <ul class="navbar-nav mr-auto mt-2 mt-lg-0">
-                        <li class="nav-item active">
-                            <a class="nav-link" href="/">Home <span class="sr-only">(current)</span></a>
-                        </li>
-                        <li class="nav-item">
-                            <a class="nav-link" href="/how-to">How to use</a>
-                        </li>
-                        <li class="nav-item"><a href="https://github.com/Mir-Khan/Play-Gen" class="nav-link">Github</a></li>
-                        <li class="nav-item"><a href="https://www.mir-khan.com/" target="_blank"
-                                class="nav-link">Resume</a></li>
-                    </ul>
-                </div>
-            </nav>
-        </div>
-        <div class="col">
-            <div class="row justify-content-center">
-                <h2>How to create a new playlist</h2>
-            </div>
-            <hr>
-            <div class="row justify-content-center list-row">
-                <ul class="list-group explain-list">
-                    <li class="list-item">
-                        <h3>Enter the name of the new playlist</h3>
-                    </li>
-                    <li class="list-item">
-                        <h3>Enter the number of songs you want in the playlist (max of 1000)</h3>
-                    </li>
-                    <li class="list-item">
-                        <h3>Decide if you want the playlist publicly visible</h3>
-                    </li>
-                    <li class="list-item">
-                        <h3>Press the submit button</h3>
-                    </li>
-                </ul>
-            </div>
-            <div class="row justify-content-center">
-                <h2>How to modify an existing playlist</h2>
-            </div>
-            <hr>
-            <div class="row justify-content-center list-row">
-                <ul class="list-group explain-list">
-                    <li class="list-item">
-                        <h3>Enter the name of the playlist you want to modify <span
-                                style="font-weight: 700; text-decoration: underline;">EXACTLY</span> as it is on Spotify
-                        </h3>
-                    </li>
-                    <li class="list-item">
-                        <h3>Enter the number of songs you want in the playlist (max of 1000)</h3>
-                    </li>
-                    <li class="list-item">
-                        <h3>Press the submit button</h3>
-                    </li>
-                </ul>
-            </div>
-            <div class="row justify-content-center">
-                <h2>Some important notes</h2>
-            </div>
-            <hr>
-            <div class="row justify-content-center list-row">
-                <ul class="list-group explain-list">
-                    <li class="list-item">
-                        <h3>When using the modify tool, make sure you have at least one song in the
-                            playlist<br> so the tool knows what to add</h3>
-                    </li>
-                    <li class="list-item">
-                        <h3>Since the creation tool depends on your listening history, it won't work well <br>for a new
-                            Spotify
-                            account</h3>
-                    </li>
-                </ul>
-            </div>
-        </div>
-        <div class="col align-self-center">
-            <footer class="footer mt-auto py-3">
-                <div class="container text-center">
-                    <p>Made by Mir Khan</p>
-                    <p><a href="/">Home</a> | <a href="/how-to">How to use</a> | <a
-						href="https://github.com/Mir-Khan/Play-Gen" target="_blank">Github</a> | <a
-						href="https://mir-khan.com" target="_blank">Resume</a></p>
-                </div>
-            </footer>
-        </div>
-    </div>
-</body>
-
-</html>
