+++ views/public/css/choose.css
@@ -1,0 +1,32 @@
+@import url("https://fonts.googleapis.com/css2?family=Dosis:wght@400;700&family=Montserrat:ital,wght@0,300;0,400;1,300;1,400&display=swap");
+.choose-row {
+  margin-top: 20rem; }
+
+.choose-button {
+  border: none;
+  background: none; }
+  .choose-button svg {
+    fill: #f9ece1;
+    width: 15rem;
+    height: 15rem; }
+  .choose-button p {
+    font-family: "Montserrat", sans-serif;
+    font-size: 1.7rem;
+    padding-top: 2rem;
+    color: #f9ece1; }
+
+@media only screen and (max-width: 410px) {
+  .choose-button svg {
+    width: 11rem;
+    height: 11rem; } }
+
+@media only screen and (max-height: 700px) {
+  .choose-row {
+    margin-top: 13rem; } }
+
+@media only screen and (max-height: 600px) {
+  .choose-row {
+    margin-top: 2rem; }
+  .choose-button svg {
+    width: 9rem;
+    height: 9rem; } }
