+++ views/public/css/error.css
@@ -1,0 +1,12 @@
+@import url("https://fonts.googleapis.com/css2?family=Dosis:wght@400;700&family=Montserrat:ital,wght@0,300;0,400;1,300;1,400&display=swap");
+.error-title {
+  color: #f9ece1;
+  font-size: 2.5rem;
+  text-decoration: underline;
+  font-family: "Dosis", sans-serif;
+  font-weight: 700; }
+
+.error-text {
+  color: #f9ece1;
+  font-size: 1.8rem;
+  font-family: "Montserrat", sans-serif; }
