+++ views/public/css/forms.css
@@ -1,0 +1,94 @@
+form {
+  margin: auto; }
+
+.form-group {
+  padding-right: 5rem; }
+
+.num-input,
+.text-input {
+  font-size: 1.2rem; }
+
+@media only screen and (max-width: 450px) {
+  .num-input,
+  .text-input {
+    font-size: 0.9rem; } }
+
+@media only screen and (max-width: 380px) {
+  .num-input,
+  .text-input {
+    font-size: 0.8rem; } }
+
+.user-forms {
+  padding-left: 5rem; }
+
+.checkbox {
+  margin-left: 2rem; }
+
+.button-div {
+  margin-top: 1.2rem; }
+
+#collapse-button:hover {
+  cursor: pointer; }
+
+#create-form {
+  margin-bottom: 2rem; }
+
+.logout {
+  padding-top: 3rem; }
+
+.feedback-text {
+  font-size: 1.2rem;
+  color: #e11809; }
+
+.tooltip-style {
+  width: 1.3rem;
+  height: 1.3rem;
+  fill: white;
+  fill-opacity: 0.6; }
+
+@media only screen and (max-width: 896px) {
+  .tooltip-style {
+    display: none !important; } }
+
+/* Removing the arrows from the number inputs, taken from W3 School*/
+/* Chrome, Safari, Edge, Opera */
+input::-webkit-outer-spin-button,
+input::-webkit-inner-spin-button {
+  -webkit-appearance: none;
+  margin: 0; }
+
+/* Firefox */
+input[type=number] {
+  -moz-appearance: textfield; }
+
+/* From W3 Schools for Autocomplete Forms*/
+.autocomplete {
+  /*the container must be positioned relative:*/
+  position: relative;
+  display: inline-block; }
+
+.autocomplete-items {
+  position: absolute;
+  border: 1px solid #d4d4d4;
+  background-color: white;
+  border-bottom: none;
+  border-top: none;
+  z-index: 99;
+  /*position the autocomplete items to be the same width as the container:*/
+  top: 100%;
+  left: 0;
+  right: 0; }
+
+.autocomplete-items div {
+  cursor: pointer;
+  background-color: #fff;
+  border-bottom: 1px solid #d4d4d4; }
+
+.autocomplete-items div:hover {
+  /*when hovering an item:*/
+  background-color: #e9e9e9; }
+
+.autocomplete-active {
+  /*when navigating through the items using the arrow keys:*/
+  background-color: DodgerBlue !important;
+  color: #ffffff; }
