+++ views/public/css/list.css
@@ -1,0 +1,17 @@
+.explain-list {
+  color: #f9ece1;
+  font-size: 1.6rem; }
+  .explain-list li::marker {
+    color: #f9ece1; }
+
+@media only screen and (max-width: 520px) {
+  .explain-list {
+    font-size: 1.5rem; } }
+
+@media only screen and (max-width: 415px) {
+  .explain-list {
+    font-size: 1.3rem; } }
+
+@media only screen and (max-width: 320px) {
+  .explain-list {
+    font-size: 0.9rem; } }
