+++ views/public/scss/style.scss
@@ -16,7 +16,7 @@ html {
     background-color: $gray;
 }
 
-h1{
+h1 {
     font-family: $titles;
     color: $linen;
     font-size: 4.5rem;
@@ -51,21 +51,10 @@ hr {
     border-color: $linen;
 }
 
-h3{
+h3 {
     color: $linen;
     font-family: $text;
     font-size: 1.6rem;
-}
-
-.explain-list{
-    li::marker{
-        color: $linen;
-    }
-}
-
-.list-row{
-    padding-bottom: 5rem;
-    margin: auto;
 }
 
 .btn {
@@ -112,136 +101,37 @@ h3{
 
 .container-fluid {
     background-color: $gray;
-    height: 100vh;
-}
-
-.nav-row {
-    background-color: $black;
-    color: $green;
-
-    nav {
-        padding-left: 2rem;
-        font-size: 1.5rem;
-
-        .navbar-brand {
-            font-family: $titles;
-            color: $green;
-            font-size: 1.5rem;
-        }
-
-        ul {
-            li {
-                // taken from https://emilkowalski.github.io/css-effects-snippets/ and slightly modified
-                display: inline-block;
-                position: relative;
-                color: hsl(222, 25%, 14%);
-
-                &:after {
-                    content: '';
-                    position: absolute;
-                    width: 100%;
-                    transform: scaleX(0);
-                    height: 2px;
-                    bottom: 0;
-                    left: 0;
-                    background-color: $vermil;
-                    transform-origin: bottom right;
-                    transition: transform 0.4s cubic-bezier(0.86, 0, 0.07, 1);
-                }
-
-                &:hover::after {
-                    transform: scaleX(1);
-                    transform-origin: bottom left;
-                }
-            }
-        }
-    }
+    height: 90vh;
 }
 
 .row {
-    margin-bottom: 2.5rem;
+    margin-bottom: 5rem;
 }
 
 .btn {
     font-size: 1.7rem;
 }
 
-.form {
-    width: 40%;
-    margin: auto;
-}
-
-.sidebar-links {
-    list-style-type: none;
-    margin: 0;
-    padding-left: 2rem;
-}
-
-.form-group {
-    padding-right: 5rem;
-}
-
-.num-input {
-    width: 90%;
-}
-
-.num-input, .text-input{
-    font-size: 1.4rem;
-}
-
-.checkbox {
-    margin-left: 2rem;
-}
-
-.button-div {
-    margin-top: 1.2rem;
-}
-
-#collapse-button:hover {
-    cursor: pointer;
-}
-
-#create-form {
-    margin-bottom: 2rem;
-}
-
-.logout{
-    padding-top: 3rem;
-}
-
-.feedback-text{
-    font-size: 1.2rem;
-    color: $vermil;
-}
-
-.footer{
+.footer {
     background-color: $gray;
-    p{
+    p {
         color: $linen;
-        font-size: 1.3rem;
+        font-size: 1.1rem;
         font-family: $text;
-        a{
+        white-space: nowrap;
+
+        a {
+            display: inline-block !important;
             text-decoration: none;
             color: $green;
-            &:hover{
+
+            &:hover {
                 color: $linen !important;
             }
-            &:visited{
+
+            &:visited {
                 color: hsl(59, 85%, 55%);
             }
         }
     }
-}
-
-/* Removing the arrows from the number inputs, taken from W3 School*/
-/* Chrome, Safari, Edge, Opera */
-input::-webkit-outer-spin-button,
-input::-webkit-inner-spin-button {
-    -webkit-appearance: none;
-    margin: 0;
-}
-
-/* Firefox */
-input[type=number] {
-    -moz-appearance: textfield;
 }
